폴더 구조 변경, 간석기 UI 적용
