+++ Assets/MakingStone.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8b53ef8f4b5ef6c419ca9fafbf3934ef
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
