+++ Assets/MakingStone/Image.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ef9b9f11b3f3f92459980140aea34754
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
