+++ Assets/MakingStone/Image/Arrow.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:06a963e9c5ef569cd3ce91ad03221ea224bd9803270f42ebff29859a61012ebb
+size 5120
