+++ Assets/MakingStone/Image/Gauge.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:79e8bfeca00be022faba63a4db153be100fbcb8e8091faf023c4e7e16d6a62c7
+size 14413
