+++ Assets/MakingStone/Script/GaugeBar.cs.meta
