+++ Assets/MakingStone/Script/StoneControl.cs.meta
